# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "UI testing for Payment Method",
        "title": "Finance Web Application",
        "contact": {},
        "version": "1.16.2"
    },
    "basePath": "/auth/",
    "paths": {
        "/register/teacher": {
            "post": {
                "security": [
                    {
                        "JWT-Token": []
                    }
                ],
                "description": "Add a new teacher to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Create a new teacher",
                "parameters": [
                    {
                        "description": "Enter the Description",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/struct_test.AdminAcc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teacher added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorModel"
                        }
                    }
                }
            }
        },
        "/secure/login": {
            "post": {
                "security": [
                    {
                        "JWT-Token": []
                    }
                ],
                "description": "Auto generate secret key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "It will generate the key that will use to log-in in swagger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully accessed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorModel"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorModel"
                        }
                    }
                }
            }
        },
        "/students": {
            "post": {
                "description": "Create a new student account with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student account",
                "parameters": [
                    {
                        "description": "Student data to be created",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/struct_test.Student"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/struct_test.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ErrorModel": {
            "type": "object",
            "properties": {
                "error": {},
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "security.SecretKey": {
            "type": "object",
            "properties": {
                "secret_key": {
                    "type": "string"
                }
            }
        },
        "struct_test.AdminAcc": {
            "description": "struct for getting the credentials",
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "struct_test.Student": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "department": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "miscellaneous": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT-Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}